feat(cli): improve and restructure item create/edit
- Split monolithic ProjectItemCreate and ProjectItemEdit modules into
  a proper submodule tree under Prix.Cli.Project.Item, with shared
  utilities in Commons.
- Extract all GraphQL queries into embedded .gql files.
- Replace `haskeline` with a custom `Zamazingo.Terminal.Prompts` module
  backed by `ansi-terminal` for improved UX
