+++ .hlint.yaml
@@ -18,6 +18,9 @@
     - { name: Data.Text, as: T }
     - { name: Data.Text.Lazy, as: TL }
     - { name: Data.Text.Encoding, as: TE }
+    - { name: Zamazingo.Base, as: Z.Base, importStyle: qualified, asRequired: true }
+    - { name: Zamazingo.Control, as: Z.Control, importStyle: qualified, asRequired: true }
+    - { name: Zamazingo.Terminal.Prompts, as: Z.Term.Prompts, importStyle: qualified, asRequired: true }
     - { name: Zamazingo.Text, as: Z.Text, importStyle: qualified, asRequired: true }
     - { name: Data.Text.IO, as: TIO }
     - { name: Options.Applicative, as: OA }
