+++ src/Prix/extras/add_assignees_to_assignable.gql
@@ -1,0 +1,7 @@
+mutation ($assignableId: ID!, $assigneeIds: [ID!]!) {
+  addAssigneesToAssignable(
+    input: { assignableId: $assignableId, assigneeIds: $assigneeIds }
+  ) {
+    clientMutationId
+  }
+}
