+++ src/Prix/extras/add_content_to_project.gql
@@ -1,0 +1,9 @@
+mutation ($projectId: ID!, $contentId: ID!) {
+  addProjectV2ItemById(
+    input: { projectId: $projectId, contentId: $contentId }
+  ) {
+    item {
+      id
+    }
+  }
+}
