+++ src/Prix/extras/lookup_repo_id.gql
@@ -1,0 +1,5 @@
+query ($owner: String!, $name: String!) {
+  repository(owner: $owner, name: $name) {
+    id
+  }
+}
