+++ src/Prix/extras/lookup_user_id.gql
@@ -1,0 +1,5 @@
+query ($login: String!) {
+  user(login: $login) {
+    id
+  }
+}
