+++ src/Prix/extras/remove_assignees_from_assignable.gql
@@ -1,0 +1,7 @@
+mutation ($assignableId: ID!, $assigneeIds: [ID!]!) {
+  removeAssigneesFromAssignable(
+    input: { assignableId: $assignableId, assigneeIds: $assigneeIds }
+  ) {
+    clientMutationId
+  }
+}
