+++ src/Prix/extras/update_issue.gql
@@ -1,0 +1,7 @@
+mutation ($issueId: ID!, $title: String!, $body: String) {
+  updateIssue(input: { id: $issueId, title: $title, body: $body }) {
+    issue {
+      id
+    }
+  }
+}
