+++ src/Prix/extras/update_project_item_field.gql
@@ -1,0 +1,19 @@
+mutation (
+  $projectId: ID!
+  $itemId: ID!
+  $fieldId: ID!
+  $value: ProjectV2FieldValue!
+) {
+  updateProjectV2ItemFieldValue(
+    input: {
+      projectId: $projectId
+      itemId: $itemId
+      fieldId: $fieldId
+      value: $value
+    }
+  ) {
+    projectV2Item {
+      id
+    }
+  }
+}
