+++ src/Prix/extras/update_pull_request.gql
@@ -1,0 +1,9 @@
+mutation ($pullRequestId: ID!, $title: String!, $body: String) {
+  updatePullRequest(
+    input: { pullRequestId: $pullRequestId, title: $title, body: $body }
+  ) {
+    pullRequest {
+      id
+    }
+  }
+}
