+++ src/Zamazingo/Base.hs
@@ -1,0 +1,5 @@
+module Zamazingo.Base where
+
+
+enumerate :: (Bounded a, Enum a) => [a]
+enumerate = [minBound .. maxBound]
