+++ src/Zamazingo/Control.hs
@@ -1,0 +1,5 @@
+module Zamazingo.Control where
+
+
+discard :: Monad m => a -> m ()
+discard _ = pure ()
